project done

+++ backend/controllers/product.controller.js
@@ -51,8 +51,12 @@ const listProducts = async (req,res) =>{
 
 const removeProduct = async (req,res) =>{
     try {   
-    
-        await productModel.findByIdAndDelete(req.body.id)
+        if(req.body.id){
+            await productModel.findByIdAndDelete(req.body.id)
+        }
+        else{
+            res.json({success:false ,message:"Something went wrong"})
+        }
 
         res.json({success:true ,message:"Product Removed"})
